feat: add model validations

+++ db/migrate/20220927032007_create_predictions.rb
@@ -3,8 +3,8 @@ class CreatePredictions < ActiveRecord::Migration[7.0]
     create_table :predictions do |t|
       t.bigint :winner_team_id
       t.bigint :loser_team_id
-      t.bigint :winner_team_goals
-      t.bigint :loser_team_goals
+      t.integer :winner_team_goals
+      t.integer :loser_team_goals
       t.timestamps
     end
     add_foreign_key :predictions, :teams, column: :winner_team_id
